1,增加转账申请 2,优化计划维护 3,增加hibernate session 的junit测试支持 4,修改用户账户详细数据结构，增加字段区别不同资金类型 5,修改奖金模式的算法 6,修改重庆时时彩 pk10 广东11选5的计划制定 7,修改玩法的最大中奖注数

+++ lottery-api/src/main/java/com/jll/game/order/OrderServiceImpl.java
@@ -246,7 +246,7 @@ public class OrderServiceImpl implements OrderService
 				walletId,				
 				wallet.getState()));
 		if(wallet != null 
-				&& wallet.getUserId() == user.getId() 
+				&& wallet.getUserId().intValue() == user.getId().intValue()
 				&& wallet.getState().intValue() == Constants.WalletState.NORMAL.getCode()) {
 			return true;
 		}
